test: BearerTokenExtractorTest 분리
